# ValueTree::callListeners optimization