--- a/modules/juce_data_structures/values/juce_ValueTree.cpp
+++ b/modules/juce_data_structures/values/juce_ValueTree.cpp
@@ -81,8 +81,8 @@ public:
             {
                 auto* v = listenersCopy.getUnchecked (i);
 
-                if (i == 0 || valueTreesWithListeners.contains (v))
-                    v->listeners.callExcluding (listenerToExclude, fn);
+                if (i == 0 || valueTreesWithListeners[i] == v || valueTreesWithListeners.contains(v))
+                    v->listeners.callExcluding(listenerToExclude, fn);
             }
         }
     }